Consolidated interactive dashboard functionality into a single file and removed deprecated learning dashboard
